backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy
pandas
yfinance
scikit-learn
tensorflow
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4